méthodes de sauvegardes

+++ TimeLog/src/timelog/Employe.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class Employe extends Personne {
 
-	public Employe(String nom, int id_personne,String poste, int taux_horaire_base, int taux_horaire_supp, Date date_embauche,
+	public Employe(String nom, int id_personne,String poste, double taux_horaire_base, double taux_horaire_supp, Date date_embauche,
 			Date date_depart, int numero_nas) {
 		super(nom, id_personne, poste, taux_horaire_base, taux_horaire_supp, date_embauche, date_depart, numero_nas);
 	}
